requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
scikit-learn
torch
optuna
pillow
ipython
